Move SeverityLevel and Severity to Models namespace
