+++ src/EventLogExpert.UI/Models/EventLogData.cs
@@ -1,7 +1,6 @@
 ﻿// // Copyright (c) Microsoft Corporation.
 // // Licensed under the MIT License.
 
-using EventLogExpert.Eventing.Helpers;
 using EventLogExpert.Eventing.Models;
 using EventLogExpert.Eventing.Readers;
